+++ src/test/java/ActualValidCalculatorTest.java
@@ -1,21 +1,46 @@
 import Entities.Locations.LocationBitboard;
-import Entities.Pieces.Pawn;
+import Entities.VariousCalculators.ActualValidCalculator;
+import Entities.VariousCalculators.Calculators;
+import Entities.VariousCalculators.CheckCalculator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-public class PawnTest{
+class ActualValidCalculatorTest {
 
-    // black pawn
-    // initial spot, no enemy
+    Calculators calculators = new Calculators();
+    ActualValidCalculator actualValidCalculator = new ActualValidCalculator(
+            calculators, new CheckCalculator(calculators));
+    LocationBitboard board = new LocationBitboard();
+
     @Test
-    public void initial_position_testb(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] array_from = {
+    void queenMovingResultsInCheck() {
+        // The array representation of queenLocation
+        int[][] whiteQueenLocation = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+
+        int[][] whiteKingLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+
+        int[][] blackRookLocation = {
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
@@ -23,190 +48,48 @@ public class PawnTest{
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        int[][] array_to = {
+
+        // Remove all pieces from board. Only have white queen on board.
+        TestHelper.remove_all_pieces(board);
+        board.whiteQueen[0] = TestHelper.bitboard_representation(whiteQueenLocation);
+        board.whiteKing[0] = TestHelper.bitboard_representation(whiteKingLocation);
+        board.blackRook[0] = TestHelper.bitboard_representation(blackRookLocation);
+        board.updateLocationVariables();
+
+        // Only moves that does not put the king in check should be valid
+        int[][] expectedActualValidMove = {
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(array_from);
-        long expected_moves = TestHelper.bitboard_representation(array_to);
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
+
+        long actualValidMove = actualValidCalculator.actual_valid_moves(
+                TestHelper.bitboard_representation(whiteQueenLocation), true, board);
+        // Check if valid move is as expected
+        assertEquals(TestHelper.bitboard_representation(expectedActualValidMove), actualValidMove);
     }
 
-    // initial spot, enemy right in front
+    // pawn movement when king would be in check
     @Test
-    public void enemy_block_testb(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] array_from = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 1, 0, 0, 0, 0, 0},
+    void pawnMovingResultsInCheck() {
+        // The array representation of pawnLocation
+        int[][] whitePawnLocation = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        int[][] enemy = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(array_from);
-        long expected_moves = 0L;
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
 
-    // not initial spot no enemy
-    @Test
-    public void random_position_testb(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // not initial spot double capture target
-    @Test
-    public void double_capture_testb() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_pawn = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 0, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // corner spot
-    @Test
-    public void border_column_capture_testb() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_pawn = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 1, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // last row
-    @Test
-    public void last_row_testb() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
+        int[][] whiteKingLocation = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
@@ -216,274 +99,8 @@ public class PawnTest{
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 1, 0, 0, 0, 0}
         };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = 0L;
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
 
-    // en passant capture
-    @Test
-    public void en_passant_capture_testb() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 1, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 1, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_pawn_from = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_pawn_to = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 1, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn_from);
-        new_board.updateLocationVariables();
-        long white_pawn_from_long = TestHelper.bitboard_representation(enemy_pawn_from);
-        long white_pawn_to_long = TestHelper.bitboard_representation(enemy_pawn_to);
-        new_board.move_piece(white_pawn_from_long, white_pawn_to_long, true);
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, false, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-
-    // white pawn
-    // initial spot, no enemy
-    @Test
-    public void initial_position_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // initial spot, enemy right in front
-    @Test
-    public void enemy_block_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = 0L;
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // not initial spot no enemy
-    @Test
-    public void random_position_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 1, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 1, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // not initial spot double capture target
-    @Test
-    public void double_capture_test() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] black_pawn = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 0, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // corner spot
-    @Test
-    public void border_column_capture_test() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] arrayto = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] black_pawn = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // last row
-    @Test
-    public void last_row_test() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
+        int[][] blackRookLocation = {
                 {0, 0, 0, 1, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
@@ -493,192 +110,163 @@ public class PawnTest{
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = 0L;
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
+
+        // Remove all pieces from board. Only have white pawn, king and black rook on board.
+        TestHelper.remove_all_pieces(board);
+        board.whitePawn[0] = TestHelper.bitboard_representation(whitePawnLocation);
+        board.whiteKing[0] = TestHelper.bitboard_representation(whiteKingLocation);
+        board.blackRook[0] = TestHelper.bitboard_representation(blackRookLocation);
+        board.updateLocationVariables();
+
+        // Only moves that does not put the king in check should be valid
+        int[][] expectedActualValidMove = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+
+        long actualValidMove = actualValidCalculator.actual_valid_moves(
+                TestHelper.bitboard_representation(whitePawnLocation), true, board);
+        // Check if valid move is as expected
+        assertEquals(TestHelper.bitboard_representation(expectedActualValidMove), actualValidMove);
     }
 
-    // en passant capture
+    // pawn movement when king would be in check, but enemy in front
     @Test
-    public void en_passant_capture_test() {
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] arrayfrom = {
+    void pawnMovingResultsInCheckBlocked() {
+        // The array representation of pawnLocation
+        int[][] whitePawnLocation = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+
+        int[][] whiteKingLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0}
+        };
+
+        int[][] blackRookLocation = {
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        int[][] arrayto = {
+        int[][] blackPawnLocation = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 1, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+
+        // Remove all pieces from board. Only have white pawn, king and black rook on board.
+        TestHelper.remove_all_pieces(board);
+        board.whitePawn[0] = TestHelper.bitboard_representation(whitePawnLocation);
+        board.whiteKing[0] = TestHelper.bitboard_representation(whiteKingLocation);
+        board.blackRook[0] = TestHelper.bitboard_representation(blackRookLocation);
+        board.blackPawn[0] = TestHelper.bitboard_representation(blackPawnLocation);
+        board.updateLocationVariables();
+
+        long actualValidMove = actualValidCalculator.actual_valid_moves(
+                TestHelper.bitboard_representation(whitePawnLocation), true, board);
+        // Check if valid move is as expected
+        assertEquals(0L, actualValidMove);
+    }
+
+    @Test
+    void KnightMovingResultInCheck() {
+        // The array representation of knightLocation
+        int[][] whiteKnightLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        int[][] black_pawn_from = {
+
+        int[][] whiteKingLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0}
+        };
+
+        int[][] blackQueenLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 1, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        int[][] black_pawn_to = {
+
+        int[][] blackRookLocation = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn_from);
-        new_board.updateLocationVariables();
-        long black_pawn_from_long = TestHelper.bitboard_representation(black_pawn_from);
-        long black_pawn_to_long = TestHelper.bitboard_representation(black_pawn_to);
-        new_board.move_piece(black_pawn_from_long, black_pawn_to_long, false);
-        long from = TestHelper.bitboard_representation(arrayfrom);
-        long expected_moves = TestHelper.bitboard_representation(arrayto);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
 
-    // attack coverage tests
-    // initial positions, no enemy
-    @Test
-    public void initial_position_attack_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] array_from = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] array_attacks = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
-        new_board.updateLocationVariables();
-        long from = TestHelper.bitboard_representation(array_from);
-        long expected_moves = TestHelper.bitboard_representation(array_attacks);
-        long test_moves = test_pawn.valid_moves(from, true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
+        // Remove all pieces from board.
+        TestHelper.remove_all_pieces(board);
+        board.whiteKnight[0] = TestHelper.bitboard_representation(whiteKnightLocation);
+        board.whiteKing[0] = TestHelper.bitboard_representation(whiteKingLocation);
+        board.blackQueen[0] = TestHelper.bitboard_representation(blackQueenLocation);
+        board.blackRook[0] = TestHelper.bitboard_representation(blackRookLocation);
+        board.updateLocationVariables();
 
-    // initial positions, 1 enemy
-    @Test
-    public void one_enemy_attack_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] ally_positions = {
+        // Only moves that does not put the king in check should be valid
+        int[][] expectedActualValidMove = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] array_attacks = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_positions = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 1, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(ally_positions);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy_positions);
-        new_board.updateLocationVariables();
-        long expected_moves = TestHelper.bitboard_representation(array_attacks);
-        long test_moves = test_pawn.attack_coverage(true, new_board);
-        assertEquals(expected_moves, test_moves);
-    }
-
-    // initial positions, multiple enemy
-    @Test
-    public void multiple_enemy_attack_test(){
-        Pawn test_pawn = new Pawn();
-        LocationBitboard new_board = new LocationBitboard();
-        TestHelper.remove_all_pieces(new_board);
-        int[][] ally_positions = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] array_attacks = {
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {1, 1, 1, 1, 1, 1, 1, 1},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0}
-        };
-        int[][] enemy_positions = {
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 1, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
-                {0, 1, 0, 1, 0, 0, 1, 0},
-                {0, 0, 0, 0, 0, 0, 0, 0},
                 {0, 0, 0, 0, 0, 0, 0, 0}
         };
-        new_board.whitePawn[0] = TestHelper.bitboard_representation(ally_positions);
-        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy_positions);
-        new_board.updateLocationVariables();
-        long expected_moves = TestHelper.bitboard_representation(array_attacks);
-        long test_moves = test_pawn.attack_coverage(true, new_board);
-        assertEquals(expected_moves, test_moves);
+
+        long actualValidMove = actualValidCalculator.actual_valid_moves(
+                TestHelper.bitboard_representation(whiteKnightLocation), true, board);
+        // Check if valid move is as expected
+        assertEquals(TestHelper.bitboard_representation(expectedActualValidMove), actualValidMove);
     }
 }
