+++ src/test/java/PawnTests.java
@@ -1,0 +1,684 @@
+import Entities.Locations.LocationBitboard;
+import Entities.Pieces.Pawn;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class PawnTests {
+
+    // black pawn
+    // initial spot, no enemy
+    @Test
+    public void initial_position_testb(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] array_from = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] array_to = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(array_from);
+        long expected_moves = TestHelper.bitboard_representation(array_to);
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // initial spot, enemy right in front
+    @Test
+    public void enemy_block_testb(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] array_from = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(array_from);
+        long expected_moves = 0L;
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // not initial spot no enemy
+    @Test
+    public void random_position_testb(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // not initial spot double capture target
+    @Test
+    public void double_capture_testb() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_pawn = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // corner spot
+    @Test
+    public void border_column_capture_testb() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_pawn = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 1, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // last row
+    @Test
+    public void last_row_testb() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = 0L;
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // en passant capture
+    @Test
+    public void en_passant_capture_testb() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 1, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 1, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_pawn_from = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_pawn_to = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 1, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(enemy_pawn_from);
+        new_board.updateLocationVariables();
+        long white_pawn_from_long = TestHelper.bitboard_representation(enemy_pawn_from);
+        long white_pawn_to_long = TestHelper.bitboard_representation(enemy_pawn_to);
+        new_board.move_piece(white_pawn_from_long, white_pawn_to_long, true);
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, false, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+
+    // white pawn
+    // initial spot, no enemy
+    @Test
+    public void initial_position_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // initial spot, enemy right in front
+    @Test
+    public void enemy_block_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = 0L;
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // not initial spot no enemy
+    @Test
+    public void random_position_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 1, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 1, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // not initial spot double capture target
+    @Test
+    public void double_capture_test() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] black_pawn = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // corner spot
+    @Test
+    public void border_column_capture_test() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] black_pawn = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // last row
+    @Test
+    public void last_row_test() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = 0L;
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // en passant capture
+    @Test
+    public void en_passant_capture_test() {
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] arrayfrom = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] arrayto = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] black_pawn_from = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] black_pawn_to = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(arrayfrom);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(black_pawn_from);
+        new_board.updateLocationVariables();
+        long black_pawn_from_long = TestHelper.bitboard_representation(black_pawn_from);
+        long black_pawn_to_long = TestHelper.bitboard_representation(black_pawn_to);
+        new_board.move_piece(black_pawn_from_long, black_pawn_to_long, false);
+        long from = TestHelper.bitboard_representation(arrayfrom);
+        long expected_moves = TestHelper.bitboard_representation(arrayto);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // attack coverage tests
+    // initial positions, no enemy
+    @Test
+    public void initial_position_attack_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] array_from = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] array_attacks = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(array_from);
+        new_board.updateLocationVariables();
+        long from = TestHelper.bitboard_representation(array_from);
+        long expected_moves = TestHelper.bitboard_representation(array_attacks);
+        long test_moves = test_pawn.valid_moves(from, true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // initial positions, 1 enemy
+    @Test
+    public void one_enemy_attack_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] ally_positions = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] array_attacks = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_positions = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 1, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(ally_positions);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy_positions);
+        new_board.updateLocationVariables();
+        long expected_moves = TestHelper.bitboard_representation(array_attacks);
+        long test_moves = test_pawn.attack_coverage(true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+
+    // initial positions, multiple enemy
+    @Test
+    public void multiple_enemy_attack_test(){
+        Pawn test_pawn = new Pawn();
+        LocationBitboard new_board = new LocationBitboard();
+        TestHelper.remove_all_pieces(new_board);
+        int[][] ally_positions = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] array_attacks = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {1, 1, 1, 1, 1, 1, 1, 1},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        int[][] enemy_positions = {
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 1, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 1, 0, 1, 0, 0, 1, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0},
+                {0, 0, 0, 0, 0, 0, 0, 0}
+        };
+        new_board.whitePawn[0] = TestHelper.bitboard_representation(ally_positions);
+        new_board.blackPawn[0] = TestHelper.bitboard_representation(enemy_positions);
+        new_board.updateLocationVariables();
+        long expected_moves = TestHelper.bitboard_representation(array_attacks);
+        long test_moves = test_pawn.attack_coverage(true, new_board);
+        assertEquals(expected_moves, test_moves);
+    }
+}
